Refactored the runners
