+++ certbot/main.py
@@ -29,6 +29,7 @@ from certbot import log
 from certbot import renewal
 from certbot import reporter
 from certbot import storage
+from certbot import updater
 from certbot import util
 
 from certbot.display import util as display_util, ops as display_ops
@@ -1075,67 +1076,7 @@ def renew_cert(config, plugins, lineage):
         notify("new certificate deployed with reload of {0} server; fullchain is {1}".format(
                config.installer, lineage.fullchain), pause=False)
         # Run deployer
-        _run_renewal_deployer(renewed_lineage, installer, config)
-
-def run_renewal_updaters(config, plugins, lineage):
-    """Run updaters that the plugin supports
-
-    :param config: Configuration object
-    :type config: interfaces.IConfig
-
-    :param plugins: List of plugins
-    :type plugins: `list` of `str`
-
-    :param lineage: Certificate lineage object
-    :type lineage: storage.RenewableCert
-
-    :returns: `None`
-    :rtype: None
-    """
-    try:
-        # installers are used in auth mode to determine domain names
-        installer, _ = plug_sel.choose_configurator_plugins(config, plugins, "certonly")
-    except errors.PluginSelectionError as e:
-        logger.info("Could not choose appropriate plugin: %s", e)
-        raise
-    _run_updaters(lineage, installer, config)
-
-def _run_renewal_deployer(lineage, installer, config):
-    """Helper function to run deployer interface method if supported by the used
-    installer plugin.
-
-    :param lineage: Certificate lineage object
-    :type lineage: storage.RenewableCert
-
-    :param installer: Installer object
-    :type installer: interfaces.IInstaller
-
-    :returns: `None`
-    :rtype: None
-    """
-    if config.installer_updates and isinstance(installer, interfaces.RenewDeployer):
-        installer.renew_deploy(lineage)
-
-def _run_updaters(lineage, installer, config):
-    """Helper function to run the updater interface methods if supported by the
-    used installer plugin.
-
-    :param lineage: Certificate lineage object
-    :type lineage: storage.RenewableCert
-
-    :param installer: Installer object
-    :type installer: interfaces.IInstaller
-
-    :returns: `None`
-    :rtype: None
-    """
-    for domain in lineage.names():
-        if config.server_tls_updates:
-            if isinstance(installer, interfaces.ServerTLSUpdater):
-                installer.server_tls_updates(domain)
-        if config.installer_updates:
-            if isinstance(installer, interfaces.GenericUpdater):
-                installer.generic_updates(domain)
+        updater.run_renewal_deployer(renewed_lineage, installer, config)
 
 def certonly(config, plugins):
     """Authenticate & obtain cert, but do not install it.
